Restructure project: move themes/fonts to styles/, i18nContext to utils/i18n/
