+++ components/views/SearchView.tsx
@@ -4,7 +4,7 @@ import { motion } from 'framer-motion';
 import { ArrowLeft, Search, X } from 'lucide-react';
 import { FileSystemItem, AppTheme } from '../../types';
 import { FileItem } from '../FileItem';
-import { useI18n } from '../../utils/i18nContext';
+import { useI18n } from '../../utils/i18n/i18nContext';
 
 interface SearchViewProps {
   items: FileSystemItem[]; // All items (non-trashed)
