+++ index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
 import './styles/glass.css';
-import './utils/fonts-imports';
+import './styles/fonts-imports';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {
